Model_Part2/requirements (3).txt
huggingface_hub==0.25.2
torch
gradio
transformers
langchain_community
langchain
sentence-transformers
faiss-cpu
langchain-huggingface
pypdf
langchain-groq
cryptography
pytesseract
pdf2image
streamlit-pdf-viewer
gradio
torch
pillow
streamlit